fix(assets)!: type Amazon Ads count metrics as integers
Every Amazon Ads count metric (impressions, clicks, purchases, units sold,
detail page views, add-to-carts, branded searches, video views, viewable
impressions, cumulative reach, and their attribution-window variants) was
declared float. A scan of all 96 Swarovski legacy tables (20.5 M rows)
shows these 63 columns hold whole numbers only, while every column with
fractional values is a rate or an amount. The schemas now say so: 362
declarations across 19 stats schemas move to `int | None`, including the
same metric family in the gross-and-invalid-traffic schemas. Kindle
royalties stays float (an amount). `campaign_applicable_budget_rule_name`
was typed float and becomes str.

Guard tests pin the split: count metrics are int in every schema, rates
and amounts stay float, and a report row serialised as `impressions: 3.0`
conforms to an integer column.

BREAKING CHANGE: existing BigQuery `amazon_ads` tables carry FLOAT64
columns for these metrics and BigQuery cannot change a column type in
place. Drop and recreate those tables (or migrate them) before the first
load with this version.

By Digitl

+++ packages/interloper-assets/src/interloper_assets/amazon_ads/schemas/brands_ad_groups_stats.py
@@ -10,11 +10,11 @@ class BrandsAdGroupsStats(Schema):
     ad_group_id: int | None = Field(..., description="The ID of the ad group")
     ad_group_name: str | None = Field(..., description="The name of the ad group")
     ad_status: str | None = Field(..., description="The status of the ad")
-    add_to_cart: float | None = Field(..., description="The number of items added to cart")
-    add_to_cart_clicks: float | None = Field(..., description="The number of clicks resulting in adding to cart")
+    add_to_cart: int | None = Field(..., description="The number of items added to cart")
+    add_to_cart_clicks: int | None = Field(..., description="The number of clicks resulting in adding to cart")
     add_to_cart_rate: float | None = Field(..., description="The rate of adding to cart")
-    branded_searches: float | None = Field(..., description="The number of branded searches")
-    branded_searches_clicks: float | None = Field(
+    branded_searches: int | None = Field(..., description="The number of branded searches")
+    branded_searches_clicks: int | None = Field(
         ..., description="The number of clicks resulting from branded searches"
     )
     campaign_budget_amount: float | None = Field(..., description="The budget amount for the campaign")
@@ -23,30 +23,30 @@ class BrandsAdGroupsStats(Schema):
     campaign_id: int | None = Field(..., description="The ID of the campaign")
     campaign_name: str | None = Field(..., description="The name of the campaign")
     campaign_status: str | None = Field(..., description="The status of the campaign")
-    clicks: float | None = Field(..., description="The number of clicks")
+    clicks: int | None = Field(..., description="The number of clicks")
     cost: float | None = Field(..., description="The cost of the ad")
     cost_type: str | None = Field(..., description="The type of cost")
     date: datetime.date | None = Field(..., description="The date of the data entry")
-    detail_page_views: float | None = Field(..., description="The number of views of the detail page")
-    detail_page_views_clicks: float | None = Field(
+    detail_page_views: int | None = Field(..., description="The number of views of the detail page")
+    detail_page_views_clicks: int | None = Field(
         ..., description="The number of clicks resulting in detail page views"
     )
     ecp_add_to_cart: float | None = Field(..., description="The ECP (effective cost per) add to cart")
-    impressions: float | None = Field(..., description="The number of impressions")
+    impressions: int | None = Field(..., description="The number of impressions")
     new_to_brand_detail_page_view_rate: float | None = Field(
         ..., description="The rate of detail page views from new-to-brand customers"
     )
-    new_to_brand_detail_page_views: float | None = Field(
+    new_to_brand_detail_page_views: int | None = Field(
         ..., description="The number of detail page views from new-to-brand customers"
     )
-    new_to_brand_detail_page_views_clicks: float | None = Field(
+    new_to_brand_detail_page_views_clicks: int | None = Field(
         ..., description="The number of clicks resulting in detail page views from new-to-brand customers"
     )
     new_to_brand_ecp_detail_page_view: float | None = Field(
         ..., description="The ECP (effective cost per) detail page view from new-to-brand customers"
     )
-    new_to_brand_purchases: float | None = Field(..., description="The number of new-to-brand purchases")
-    new_to_brand_purchases_clicks: float | None = Field(
+    new_to_brand_purchases: int | None = Field(..., description="The number of new-to-brand purchases")
+    new_to_brand_purchases_clicks: int | None = Field(
         ..., description="The number of clicks resulting in new-to-brand purchases"
     )
     new_to_brand_purchases_percentage: float | None = Field(
@@ -62,28 +62,28 @@ class BrandsAdGroupsStats(Schema):
     new_to_brand_sales_percentage: float | None = Field(
         ..., description="The percentage of sales from new-to-brand customers"
     )
-    new_to_brand_units_sold: float | None = Field(
+    new_to_brand_units_sold: int | None = Field(
         ..., description="The number of units sold from new-to-brand customers"
     )
-    new_to_brand_units_sold_clicks: float | None = Field(
+    new_to_brand_units_sold_clicks: int | None = Field(
         ..., description="The number of clicks resulting in units sold from new-to-brand customers"
     )
     new_to_brand_units_sold_percentage: float | None = Field(
         ..., description="The percentage of units sold from new-to-brand customers"
     )
-    purchases: float | None = Field(..., description="The number of purchases")
-    purchases_clicks: float | None = Field(..., description="The number of clicks resulting in purchases")
-    purchases_promoted: float | None = Field(..., description="The number of promoted purchases")
+    purchases: int | None = Field(..., description="The number of purchases")
+    purchases_clicks: int | None = Field(..., description="The number of clicks resulting in purchases")
+    purchases_promoted: int | None = Field(..., description="The number of promoted purchases")
     sales: float | None = Field(..., description="The total sales amount")
     sales_clicks: float | None = Field(..., description="The number of sales clicks")
     sales_promoted: float | None = Field(..., description="The total promoted sales amount")
-    units_sold: float | None = Field(..., description="The number of units sold")
-    units_sold_clicks: float | None = Field(..., description="The number of clicks resulting in units sold")
+    units_sold: int | None = Field(..., description="The number of units sold")
+    units_sold_clicks: int | None = Field(..., description="The number of clicks resulting in units sold")
     video_5_second_view_rate: float | None = Field(..., description="The rate of video 5-second views")
-    video_5_second_views: float | None = Field(..., description="The number of video 5-second views")
-    video_complete_views: float | None = Field(..., description="The number of complete video views")
-    video_first_quartile_views: float | None = Field(..., description="The number of video first quartile views")
-    video_midpoint_views: float | None = Field(..., description="The number of video midpoint views")
-    video_third_quartile_views: float | None = Field(..., description="The number of video third quartile views")
-    video_unmutes: float | None = Field(..., description="The number of video unmutes")
+    video_5_second_views: int | None = Field(..., description="The number of video 5-second views")
+    video_complete_views: int | None = Field(..., description="The number of complete video views")
+    video_first_quartile_views: int | None = Field(..., description="The number of video first quartile views")
+    video_midpoint_views: int | None = Field(..., description="The number of video midpoint views")
+    video_third_quartile_views: int | None = Field(..., description="The number of video third quartile views")
+    video_unmutes: int | None = Field(..., description="The number of video unmutes")
     viewability_rate: float | None = Field(..., description="The rate of viewability for the ad")
